backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
transformers==4.38.2
requests==2.31.0
python-dotenv==1.0.1